backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.0"
description = ""
authors = ["diogo464 <diogo464@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
minio = "^7.1.13"
nats-py = "^2.2.0"
pydantic = "^1.10.4"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
SQLAlchemy = "^2.0.1"
opentelemetry-proto = "^1.15.0"
redis = {extras = ["hiredis"], version = "^4.4.2"}
geoip2 = "^4.6.0"
rocketry = "^2.5.1"
psycopg = {extras = ["binary"], version = "^3.1.8"}

[tool.poetry.dev-dependencies]
black = "^23.1.0"
mypy = "^0.991"
types-requests = "^2.28.11"
types-redis = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
